// apps/web/src/lib/api/generated/model/taskOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface TaskOut {
  assignee_user_id: string;
  completed_at: string | null;
  created_at: string;
  created_by: string | null;
  due_at: string | null;
  entity_id: string | null;
  entity_type: string | null;
  id: string;
  notes: string | null;
  priority: string;
  status: string;
  title: string;
  version: number;
}